+++ CHANGELOG.md
@@ -2,6 +2,36 @@ Release History
 ---------------
 
 ## [Unreleased]
+# [1.4.2] - 2016-09-06
+
+### Fixed
+- error when exporting hidden columns to excel 
+
+# [1.4.1] - 2016-08-30
+
+### Fixed
+- null values in data were rendered as "null" in IE, now they're converted to empty string
+
+# [1.4.0] - 2016-08-30
+### Breaking changes
+- Expand property in field settings is no longer used to define which directive to use when expanding/opening a row, just pass true if clicking column should expand row. Use gtExpand attribute to pass and object containing which directive to use and optionally, if multiple rows should be allowed, add that property as well like this:
+`expandConfig:{
+    directive:'<my-custom-directive></my-custom-directive>',
+    multiple:true // false by default
+ }`
+
+### Added
+- Support for responsive layout using stacked columns
+- Support for custom sort function
+- Support for only having one row expanded/opened at a time 
+
+### Fixed
+- Sorting of null vales
+
+### Improved
+- Removed some unnecessary two-way-bindings and watches
+- Scopes created using $compile are now removed along with their watches (if any)
+
 # [1.3.3] - 2016-08-19
 ### Fixed
 - Compile error